Handle non-JSON response from Salesforce should
This is a simple fix for responses from Salesforce that are not properly formatted JSON.  An example could be an APEX error from a custom APEX rest api (depending on how that is written).  Although this `catch` shouldn't get exercised, this will prevent an app from cashing in case it does.